Reorganize mock map package
