+++ felix/.gitignore
@@ -22,6 +22,8 @@
 /artifacts/
 /bpf-gpl/include/libbpf/src/amd64/
 /bpf-gpl/include/libbpf/src/arm64/
+bpf-gpl/include/libbpf/src/armv7/
+bpf-gpl/include/libbpf/src/ppc64le/
 
 # IDE files.
 /.idea/
